requirements.txt
visdom
tqdm
pillow
numpy
